% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pvalue}
\alias{pvalue}
\title{Function to Compute P-value.}
\usage{
pvalue(t, df = Inf, alternative = c("two.sided", "less", "greater"))
}
\arguments{
\item{t}{t-statistic.}

\item{df}{degree of freedoms.}

\item{alternative}{type of alternatives.}
}
\description{
Function to compute p-value based on a t-statistic.
}
\examples{
pvalue(1.96, alternative = "greater")
}
\keyword{misc}

