% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zim.R
\name{zim.fit}
\alias{zim.fit}
\title{Fitter Function for Zero-Inflated Models}
\usage{
zim.fit(y, X, Z, weights = rep(1, nobs), offset = rep(0, nobs),
  control = zim.control(...), ...)
}
\arguments{
\item{y}{response variable.}

\item{X}{design matrix for log-linear part.}

\item{Z}{design matrix for logistic part.}

\item{weights}{an optional vector of 'prior weights' to be used in the fitting process.}

\item{offset}{offset variable}

\item{control}{control arguments from \code{\link{zim.control}}.}

\item{...}{additional argumetns.}
}
\description{
\code{\link{zim.fit}} is the basic computing engine called by \code{\link{zim}} used to fit 
zero-inflated models. This should usually \emph{not} be used directly unless by experienced users.
}
\seealso{
\code{\link{zim}}, \code{\link{zim.control}}
}
\keyword{regression}
