% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dzim.R
\name{dzim.plot}
\alias{dzim.plot}
\title{Trace Plots from DZIM}
\usage{
dzim.plot(object, k.inv = FALSE, sigma.sq = FALSE, ...)
}
\arguments{
\item{object}{objective from \code{\link{dzim}} or \code{\link{dzim.fit}}.}

\item{k.inv}{logical; indicating whether an inverse transformation is needed for the dispersion parameter.}

\item{sigma.sq}{logical; indicating whether a square transformation is needed for the standard deviation parameter.}

\item{...}{additional arguments.}
}
\description{
Function to display trace plots from a dynamic zero-inflated model.
}
